requirements.txt
python-binance>=1.0.17
python-dotenv>=1.0.0
typer>=0.9.0
rich>=13.0.0
